refactor: de-part production_panel cluster (#4117)
Migrate the 7-part production panel to explicit-import libraries,
merge extension parts into their subpanel modules, and shrink the
no-part grandfather allowlist by 8 entries.
