--- app/lib/features/game/widgets/production/production_panel_support_allocation_summary.dart
@@ -1,63 +1,0 @@
-/// Allocation subpanel header and labour summary for production screen.
-/// SPEC/ui/production-panel.md.
-
-part of 'production_panel.dart';
-
-extension _AllocationSubpanelSummary on _AllocationSubpanel {
-  int computeTotalRequiredLabour() {
-    return desiredOutputByRecipe.entries.fold<int>(0, (sum, entry) {
-      final recipe = ProductionRecipesCatalog.byId[entry.key];
-      if (recipe == null) {
-        return sum;
-      }
-      return sum + entry.value * recipe.labourPerOutput;
-    });
-  }
-
-  Widget buildAllocationHeader(ThemeData theme) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        Expanded(
-          child: Text(
-            l10n.production_allocation,
-            style: theme.textTheme.titleSmall,
-          ),
-        ),
-        CtDangerTextButton(
-          key: const ValueKey<String>('production_allocation_reset_button'),
-          onPressed: () => onDesiredOutputChanged({}),
-          label: l10n.common_reset,
-          tooltip: l10n.common_reset,
-        ),
-      ],
-    );
-  }
-
-  List<Widget> buildLabourSummary(
-    ThemeData theme,
-    int totalRequiredLabour,
-    bool labourInsufficient,
-  ) {
-    return <Widget>[
-      CtGap.m,
-      Text(
-        l10n.production_totalLabour(totalRequiredLabour, effectiveLabour),
-        style: theme.textTheme.bodySmall?.copyWith(
-          fontWeight: FontWeight.w600,
-          color: labourInsufficient ? theme.colorScheme.error : null,
-        ),
-      ),
-      if (labourInsufficient)
-        Padding(
-          padding: const EdgeInsets.only(top: 4),
-          child: Text(
-            l10n.production_labourInsufficient,
-            style: theme.textTheme.bodySmall?.copyWith(
-              color: theme.colorScheme.error,
-            ),
-          ),
-        ),
-    ];
-  }
-}
